Merlixo support tfe cloud

+++ main.go
@@ -3,6 +3,7 @@ package main
 import (
 	"context"
 	_ "embed"
+	"encoding/json"
 	"flag"
 	"fmt"
 	"os"
@@ -14,6 +15,8 @@ import (
 	"github.com/busser/tfautomv/internal/terraform"
 	"github.com/busser/tfautomv/internal/tfautomv"
 	"github.com/busser/tfautomv/internal/tfautomv/ignore"
+
+	tfjson "github.com/hashicorp/terraform-json"
 )
 
 func main() {
@@ -78,26 +81,42 @@ func run() error {
 	// Terraform's plan contains a lot of information. For now, this is all we
 	// need. In the future, we may choose to use other sources of information.
 
-	logln("Running \"terraform init\"...")
-	err = tf.Init(ctx)
-	if err != nil {
-		return err
+	var plan *tfjson.Plan
+
+	if jsonPlanFile == "" {
+
+		logln("Running \"terraform init\"...")
+		err = tf.Init(ctx)
+		if err != nil {
+			return err
+		}
+		
+		logln("Running \"terraform plan\"...")
+		planFile, err := os.CreateTemp("", "tfautomv.*.plan")
+		if err != nil {
+			return err
+		}
+		defer os.Remove(planFile.Name())
+
+		if _, err := tf.Plan(context.TODO(), tfexec.Out(planFile.Name())); err != nil {
+			return err
+		}
+		plan, err = tf.ShowPlanFile(context.TODO(), planFile.Name())
+		if err != nil {
+			return err
+		}
+
+	} else {
+		jsonPlan, err := os.ReadFile(jsonPlanFile)
+		if err != nil {
+			return err
+		}
+		if err := json.Unmarshal(jsonPlan, &plan); err != nil {
+			return err
+		}
 	}
 
-	logln("Running \"terraform plan\"...")
-	planFile, err := os.CreateTemp("", "tfautomv.*.plan")
-	if err != nil {
-		return err
-	}
-	defer os.Remove(planFile.Name())
-	if _, err := tf.Plan(context.TODO(), tfexec.Out(planFile.Name())); err != nil {
-		return err
-	}
-	plan, err := tf.ShowPlanFile(context.TODO(), planFile.Name())
-	if err != nil {
-		return err
-	}
-
+	logln("Analysing plan...")
 	analysis, err := tfautomv.AnalysisFromPlan(plan, rules)
 	if err != nil {
 		return err
@@ -154,6 +173,7 @@ var (
 	printVersion bool
 	showAnalysis bool
 	terraformBin string
+	jsonPlanFile string
 )
 
 func parseFlags() {
@@ -164,6 +184,7 @@ func parseFlags() {
 	flag.BoolVar(&showAnalysis, "show-analysis", false, "show detailed analysis of Terraform plan")
 	flag.BoolVar(&printVersion, "version", false, "print version and exit")
 	flag.StringVar(&terraformBin, "terraform-bin", "terraform", "terraform binary to use")
+	flag.StringVar(&jsonPlanFile, "plan-file", "", "local file with Plan in json format")
 
 	flag.Parse()
 }
